pyproject.toml
[tool.poetry]
name = "_fastapi_"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
pydantic = "^1.10.2"
uvicorn = "^0.18.3"
yoyo-migrations = "^7.3.2"
dependency-injector = "^4.40.0"
aiocache = "^0.11.1"
aiopg = "^1.3.4"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
